Refactored authorization to main process
