+++ ExilenceNextApp/public/main/authWindow.js
@@ -1,0 +1,50 @@
+const { ipcMain, BrowserWindow } = require('electron');
+
+const checkForMissingWindow = require('../util');
+
+const createAuthWindow = ({ _event, mainWindow, options }) => {
+  checkForMissingWindow({ category: 'authWindow', mainWindow });
+
+  let authWindow = new BrowserWindow({
+    width: 500,
+    height: 750,
+    show: false,
+    autoHideMenuBar: true,
+    nodeIntegration: true,
+    fullscreenable: false,
+    resizable: false,
+    minimizable: false,
+    maximizable: false,
+    alwaysOnTop: true,
+  });
+
+  const authUrl = `https://www.pathofexile.com/oauth/authorize?client_id=${options.clientId}&response_type=${options.responseType}&scope=${options.scopes}&state=${options.state}&redirect_uri=${options.redirectUrl}`;
+
+  authWindow.webContents.on('will-redirect', (_event, url) => {
+    // todo: handle returned state
+    const raw_code = /code=([^&]*)/.exec(url) || null;
+    const code = raw_code && raw_code.length > 1 ? raw_code[1] : null;
+    const error = /\?error=(.+)$/.exec(url);
+
+    if (code || error) {
+      authWindow.destroy();
+    }
+    mainWindow.webContents.send('auth-callback', { code, error });
+  });
+
+  // Reset the authWindow on close
+  authWindow.on(
+    'close',
+    function () {
+      authWindow = null;
+    },
+    false
+  );
+
+  authWindow.loadURL(authUrl);
+  authWindow.show();
+};
+
+module.exports = {
+  createAuthWindow,
+};
