+++ ExilenceNextApp/public/main/index.js
@@ -1,6 +1,7 @@
 exports.autoUpdater = require('./autoUpdater');
 exports.flashFrame = require('./flashFrame');
 exports.logMonitor = require('./logMonitor');
+exports.authWindow = require('./authWindow');
 exports.tray = require('./tray');
 exports.netWorthOverlay = require('./overlays/netWorthOverlay');
 exports.menuFunctions = require('./menuFunctions');
